requirements.txt
Flask==2.2.2
requests==2.28.1
Werkzeug==2.2.2
google-generativeai==0.4.0
python-dotenv==0.21.0
Flask-Cors==3.0.10
